+++ database/migrations/2019_02_20_100249_create_modules_table.php
@@ -14,8 +14,8 @@ class CreateModulesTable extends Migration
     public function up()
     {
         Schema::create('modules', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('user_id')->unsigned();
+            $table->bigIncrements('id');
+            $table->bigInteger('user_id')->unsigned();
             $table->string('title');
             $table->string('subtitle')->nullable();
             $table->text('text')->nullable();
